+++ src/main/java/com/space/server/domain/ai/service/dto/request/gpt/AiChat.java
