+++ src/main/java/com/space/server/domain/ai/service/dto/response/AiResponse.java
