+++ src/main/java/com/space/server/domain/ai/service/implementation/PromptCreator.java
