+++ src/main/java/com/space/server/domain/chat/domain/repository/ChatRepository.java
@@ -1,7 +1,7 @@
 package com.space.server.domain.chat.domain.repository;
 
 import com.space.server.domain.chat.domain.Chat;
-import com.space.server.state.domain.State;
+import com.space.server.domain.state.domain.State;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
