+++ src/main/java/com/space/server/domain/chat/service/CommandChatService.java
@@ -21,11 +21,11 @@ import com.space.server.domain.checklist.domain.Checklist;
 import com.space.server.domain.checklist.service.implementation.ChecklistReader;
 import com.space.server.domain.quiz.domain.Quiz;
 import com.space.server.domain.quiz.service.implementation.QuizReader;
-import com.space.server.state.domain.State;
-import com.space.server.state.domain.value.Status;
-import com.space.server.state.service.implementation.StateCreator;
-import com.space.server.state.service.implementation.StateReader;
-import com.space.server.state.service.implementation.StateUpdater;
+import com.space.server.domain.state.domain.State;
+import com.space.server.domain.state.domain.value.Status;
+import com.space.server.domain.state.service.implementation.StateCreator;
+import com.space.server.domain.state.service.implementation.StateReader;
+import com.space.server.domain.state.service.implementation.StateUpdater;
 import com.space.server.user.domain.Users;
 import com.space.server.user.service.implementation.UserReader;
 import lombok.RequiredArgsConstructor;
