+++ src/main/java/com/space/server/domain/chat/service/implementation/ChatReader.java
@@ -2,7 +2,7 @@ package com.space.server.domain.chat.service.implementation;
 
 import com.space.server.domain.chat.domain.Chat;
 import com.space.server.domain.chat.domain.repository.ChatRepository;
-import com.space.server.state.domain.State;
+import com.space.server.domain.state.domain.State;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
