+++ src/main/java/com/space/server/domain/state/domain/State.java
@@ -1,7 +1,7 @@
-package com.space.server.state.domain;
+package com.space.server.domain.state.domain;
 
 import com.space.server.domain.quiz.domain.Quiz;
-import com.space.server.state.domain.value.Status;
+import com.space.server.domain.state.domain.value.Status;
 import com.space.server.user.domain.Users;
 import io.hypersistence.utils.hibernate.type.array.IntArrayType;
 import io.hypersistence.utils.hibernate.type.array.StringArrayType;
