+++ src/main/java/com/space/server/domain/state/domain/repository/StateRepository.java
@@ -1,7 +1,7 @@
-package com.space.server.state.domain.repository;
+package com.space.server.domain.state.domain.repository;
 
 import com.space.server.domain.quiz.domain.Quiz;
-import com.space.server.state.domain.State;
+import com.space.server.domain.state.domain.State;
 import com.space.server.user.domain.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
