+++ src/main/java/com/space/server/domain/state/domain/value/Status.java
@@ -1,0 +1,6 @@
+package com.space.server.domain.state.domain.value;
+
+public enum Status {
+    SUCCESS,
+    FAIL
+}
