+++ src/main/java/com/space/server/domain/state/service/implementation/StateDeleter.java
@@ -1,7 +1,7 @@
-package com.space.server.state.service.implementation;
+package com.space.server.domain.state.service.implementation;
 
-import com.space.server.state.domain.State;
-import com.space.server.state.domain.repository.StateRepository;
+import com.space.server.domain.state.domain.State;
+import com.space.server.domain.state.domain.repository.StateRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
