+++ src/main/java/com/space/server/domain/state/service/implementation/StateReader.java
@@ -1,9 +1,9 @@
-package com.space.server.state.service.implementation;
+package com.space.server.domain.state.service.implementation;
 
 import com.space.server.domain.quiz.domain.Quiz;
-import com.space.server.state.domain.State;
-import com.space.server.state.domain.repository.StateRepository;
-import com.space.server.state.exception.StateNotFoundException;
+import com.space.server.domain.state.domain.repository.StateRepository;
+import com.space.server.domain.state.domain.State;
+import com.space.server.domain.state.exception.StateNotFoundException;
 import com.space.server.user.domain.Users;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
