--- src/main/java/com/space/server/state/domain/value/Status.java
@@ -1,6 +1,0 @@
-package com.space.server.state.domain.value;
-
-public enum Status {
-    SUCCESS,
-    FAIL
-}
